// src/types/homeTemplist.ts
/*
 * @message: 描述
 * @Author: Roy
 * @Email: @163.com
 * @Github: @163.com
 * @Date: 2021-07-12 22:42:24
 * @LastEditors: Roy
 * @LastEditTime: 2021-07-12 22:42:38
 * @Deprecated: 否
 * @FilePath: /cybertron-designer/src/types/homeTemplist.ts
 */
export interface TemplateProps {
  id: number;
  coverImg: string;
  title: string;
  author: string;
  copiedCount: number;
}